feat. 아키텍처 구성 수정
