+++ src/main/java/com/special_lecture/api/application/dto/request/LectureApplicationRequest.java
@@ -1,0 +1,22 @@
+package com.special_lecture.api.application.dto.request;
+
+import com.special_lecture.api.business.model.dto.LectureCommand;
+import com.special_lecture.common.mapper.DtoConverter;
+
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Positive;
+import lombok.Data;
+
+@Data
+public class LectureApplicationRequest {
+
+    @NotNull(message = "User ID is required.")
+    private Long userId;
+
+    @NotNull(message = "Lecture ID is required.")
+    private Long lectureId;
+
+    public LectureCommand toCommand() {
+        return DtoConverter.convert(this, LectureCommand.class);
+    }
+}
