+++ src/main/java/com/special_lecture/api/business/repo/LectureApplicationRepository.java
@@ -1,0 +1,13 @@
+package com.special_lecture.api.business.repo;
+
+import java.util.Optional;
+
+import com.special_lecture.api.business.model.entity.LectureApplication;
+
+public interface LectureApplicationRepository {
+	Optional<LectureApplication> findByUserIdAndLectureId(Long userId, Long lectureId);
+
+	int countByLectureId(Long lectureId);
+
+	void saveWithOptimisticLock(LectureApplication application);
+}
