+++ src/main/java/com/special_lecture/api/business/repo/LectureRepository.java
@@ -1,0 +1,9 @@
+package com.special_lecture.api.business.repo;
+
+import java.util.Optional;
+
+import com.special_lecture.api.business.model.entity.LectureApplication;
+
+public interface LectureRepository {
+
+}
