+++ src/main/java/com/special_lecture/common/mapper/DtoConverter.java
@@ -1,0 +1,11 @@
+package com.special_lecture.common.mapper;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+
+public class DtoConverter {
+    private static final ObjectMapper objectMapper = new ObjectMapper();
+
+    public static <T> T convert(Object source, Class<T> targetClass) {
+        return objectMapper.convertValue(source, targetClass);
+    }
+}
