--- src/main/java/com/special_lecture/domain/dto/request/LectureApplicationRequest.java
@@ -1,9 +1,0 @@
-package com.special_lecture.domain.dto.request;
-
-import lombok.Data;
-
-@Data
-public class LectureApplicationRequest {
-    private final Long userId;
-    private final Long lectureId;
-}
