--- src/main/java/com/special_lecture/domain/dto/response/LectureApplicationResponse.java
@@ -1,8 +1,0 @@
-package com.special_lecture.domain.dto.response;
-
-import lombok.Data;
-
-@Data
-public class LectureApplicationResponse {
-    private final boolean success;
-}
