--- src/main/java/com/special_lecture/domain/type/LectureStatus.java
@@ -1,7 +1,0 @@
-package com.special_lecture.domain.type;
-
-public enum LectureStatus {
-	COMPLETED,
-	PENDING, // 30안에 못들었지만, 취소건이 생기면 연락할 대기중인 예비 수강생
-	CANCELLED
-}
