--- src/main/java/com/special_lecture/repository/LectureRepository.java
@@ -1,10 +1,0 @@
-package com.special_lecture.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import com.special_lecture.domain.entity.Lecture;
-
-@Repository
-public interface LectureRepository extends JpaRepository<Lecture, Long> {
-}
